+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v3.5.0-beta2] - Unreleased
+### Fixed
+- `Font.charsPadding` loading
+### Removed
+- `DroppedFiles` class replaced with `std::vector<std::string> raylib::GetDroppedFiles()`
+
 ## [v3.5.0-beta1] - 2021-01-02
 ### Added
 - `Model.Draw()` and `Model.DrawWires()`